% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{expand_attributes}
\alias{expand_attributes}
\title{Extracts key=value; pair information in the comment column}
\usage{
expand_attributes(tabseq, output_wide = TRUE, include_sequence = TRUE, ...)
}
\arguments{
\item{tabseq}{A tabseq tibble}

\item{output_wide}{Disables the internal pivot_wide call. Defaults to TRUE. By setting this to FALSE you get a longer table instead, with is also computed faster.}

\item{include_sequence}{Includes the sequence column in the return value. Defaults to TRUE. By setting this to FALSE you will get a table without sequences, which is also computed faster.}
}
\value{
A tabseq tibble
}
\description{
This function expands tabseq comment key-value pair contents into several columns
}
