% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabseq.r
\name{extract_attributes}
\alias{extract_attributes}
\title{Extracts key=value; pair information in the comment column}
\usage{
extract_attributes(tabseq, output_wide = TRUE, include_sequence = TRUE, ...)
}
\arguments{
\item{tabseq}{A tabseq tibble}
}
\value{
A tabseq tibble
}
\description{
This function expands tabseq comment key-value pair contents into several columns
}
