% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabseq.r
\name{GC_content}
\alias{GC_content}
\title{Calculate GC content}
\usage{
GC_content(string, position = 0)
}
\arguments{
\item{string}{A string of nucleotides}

\item{position}{The position to look for GC at. 0 (default) means all, 1 means GC1, 2 means GC2 and 3 means GC3.}
}
\value{
A list of two items: annotation and fasta. The fasta item is read with tabseq::read_fasta()
}
\description{
Calculates GC, GC1 GC2 or GC3 for a given sequence.
}
