% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabseq.r
\name{read_fasta}
\alias{read_fasta}
\title{Read a fasta file and convert it to tabseq}
\usage{
read_fasta(file, basename_only = T, skip = 0)
}
\arguments{
\item{file}{The path to a fasta-file}

\item{basename_only}{If TRUE, the full paths to the files will be truncated in favor of the basename only}

\item{skip}{Skip this number of lines in the file before reading}
}
\value{
A tibble containing the respective records and sequences from the fasta-file. Represented in the tabseq format.
}
\description{
Takes the path to a fasta file. Loads the fasta file, and converts it to a tabseq table.
}
\examples{

#my_sequences = read_tabseq("path/to/sequences.fa")

}
