% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_gff}
\alias{read_gff}
\title{Read a GFF3 file}
\usage{
read_gff(file, parse_attributes = TRUE)
}
\arguments{
\item{file}{A GFF3 file}

\item{parse_attributes}{Whether the \code{attributes} column in the gff should be parsed into separate columns}
}
\value{
A list of two items: annotation and fasta. The fasta item is read with tabseq::read_fasta()
}
\description{
Read a GFF3 file
}
