% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabseq.r
\name{read_tabseq}
\alias{read_tabseq}
\title{Read a fasta file and convert it to tabseq}
\usage{
read_tabseq(file, from_fasta = T)
}
\arguments{
\item{file}{The path to a fasta-file}

\item{from_fasta}{A boolean designating whether the file path points to a fasta-file}
}
\value{
A tibble containing the respective records and sequences from the fasta-file. Represented in the tabseq format.
}
\description{
Takes the path to a fasta file. Loads the fasta file, and converts it to a tabseq table.
}
\examples{

#my_sequences = read_tabseq("path/to/sequences.fa")

}
