% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Complement nucleotide letters}
\usage{
reverse_complement(string, reverse = TRUE, type = "nucleotide")
}
\arguments{
\item{string}{A string of DNA i.e. "ACGTTTN"}

\item{reverse}{Whether the string should be reversed or not (default TRUE)}

\item{type}{String type. (default "nucleotide")}
}
\value{
A string which represents the chemical complements of the given input string
}
\description{
Takes a string of IUPAC nucleotide letters and complements them.
}
