% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{ts_neutralize}
\alias{ts_neutralize}
\title{Simplify the sequence column for development purposes}
\usage{
ts_neutralize(input)
}
\arguments{
\item{input}{A string of nucleotides}
}
\value{
A list of two items: annotation and fasta. The fasta item is read with tabseq::read_fasta()
}
\description{
Truncates the sequence column
}
