% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabseq.r
\name{tsView}
\alias{tsView}
\title{Invoke View without the sequence column}
\usage{
tsView(input)
}
\arguments{
\item{input}{A tabseq table}
}
\value{
A list of two items: annotation and fasta. The fasta item is read with tabseq::read_fasta()
}
\description{
Calculates GC, GC1 GC2 or GC3 for a given sequence.
}
