% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{write_fasta}
\alias{write_fasta}
\title{Write a tabseq table to disk in fasta-format.}
\usage{
write_fasta(x, file, format = "fasta", record_format = "\%part", verbose = F)
}
\arguments{
\item{x}{The tabseq table to save}

\item{file}{The path to a wanted fasta-file}

\item{format}{The format you want to save the file as}

\item{record_format}{A string designating how to encode the records in the fasta file written. Defaults to "\\%part", but may be a combination of all metadata columns including custom separators i.e. "\\%sample|\\%part|\\%comment.}

\item{verbose}{For debugging}
}
\value{
Saves the content to disk, and returns a tibble which represents what is being written to disk.
}
\description{
Takes a tabseq table, and saves it to disk in the fasta format.
}
\examples{

#my_sequences = read_tabseq("path/to/sequences.fa")
#write_tabseq(my_sequences, "write/sequences/here/sequences.fa")
}
