% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_tabseq}
\alias{read_tabseq}
\title{Convenience function for reading tabseqs}
\usage{
read_tabseq(input)
}
\arguments{
\item{input}{A tabseq path/filename}
}
\value{
A tibble containing the input tabseq content.
}
\description{
Imports a tabseq with the correct column types
}
