% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabseq.r
\name{write_tabseq}
\alias{write_tabseq}
\title{Write a tabseq table to disk in fasta-format.}
\usage{
write_tabseq(x, file, record_format = "\%part", to_fasta = T)
}
\arguments{
\item{x}{The tabseq table to save}

\item{file}{The path to a wanted fasta-file}

\item{record_format}{nopercA string designating how to encode the records in the fasta file. Defaults to "\%part", but may be a combination of all metadata columns including custom separators i.e. "\%sample|\%part|\%comment.}

\item{to_fasta}{dummy boolean what}
}
\value{
Saves the content to disk, and returns the tibble before it is converted to a raw string.
}
\description{
Takes a tabseq table, and saves it to disk in the fasta-format.
}
\examples{

#my_sequences = tabseq::read_tabseq(file = "extdata/test2.fa")
#tabseq::write_tabseq(my_sequences, "extdata/output.fasta")
}
